+++ docs/apps/schedy/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 * Added new result marker ``OVERLAY_REVERT_ON_NO_RESULT`` to cancel an
   overlay when the schedule produces no result.
+* Result markers can now be added by postprocessors as well.
 
 ### Changed
 * The wanted value of a room is not sent to actors at startup when